ref: move audio-icon to icons folder
