+++ resources/js/components/icons/AudioIcon.vue
